21/11/2023
Changes made:
Fixed card issues as they where moving side to side. realized that I had added padding. I fixed it by changing it to margin  4 as well as placing the cards into a container. Changed the navbar layout to be able to fit more elements. Played around with the cards layout and found a style that I liked.