Create README.md
ShopBy is a PHP web application that offers consumers an online platform to browse and buy different alcohol goods. The system is made up of a number of PHP scripts that may be used to register accounts, modify account information, add and remove products from the application, and list quantity amounts. I created a page on alcohol that would display the things that customers may buy. In order to allow users to add items to their carts and then change the amount they add, I also included a button for further information. After doing those actions, I was able to include an admin login tab. The admin would be able to add and remove items after logging in.